Cleanup of components to make things more sane and publishable
