+++ webui/packages/eidolon-components/src/lib/util.ts
@@ -1,0 +1,11 @@
+export interface CopilotParams {
+  "agent": string,
+  "operation": string,
+  "inputLabel": string,
+  "titleOperationName": string | undefined,
+  "allowSpeech": boolean,
+  "speechAgent": string | undefined,
+  "speechOperation": string | undefined
+}
+export interface DevParams {
+}
